Addressed all but Professors last two points
