+++ src/App.jsx
@@ -1,11 +1,6 @@
-import { useState } from 'react';
-import logo from './logo.svg';
-import './App.css';
-import { HomePage } from './HomePage';
+import { HomePage } from './pages/HomePage';
 
 const App = () => {
-  const [count, setCount] = useState(0);
-
   return <HomePage />;
 };
