--- src/styles.css
@@ -1,242 +1,0 @@
-header {
-    box-sizing: border-box;
-    top: 0;
-    left: 0;
-    position: fixed;
-    width: 100%;
-    background-color: white;
-    font-family: sans-serif;
-}
-
-header * {
-    display:flex;
-    height: 100%;
-}
-
-.newleft {
-    width: 75%;
-    padding-left: 2%;
-    font-family: 'Comfortaa', cursive;
-}
-
-.topright  {
-    width: 100%;
-    padding-right: 2%;
-    justify-content: flex-end;
-    font-size: 65%;
-    padding-top: 1%;
-}
-
-.menu {
-    color: blue;
-    padding-left: 2%;
-    padding-top: 13px;
-    font-size: medium;
-}
-
-body {
-    background-color: rgb(233, 232, 231);
-    padding: 2%;
-    padding-top: 90px;
-    font-family: sans-serif;
-}
-
-section {
-    box-sizing: border-box;
-    margin-bottom: 15px;
-}
-
-section *{
-    /* border: solid 1px #000; */
-    display: flex;   
-}
-
-.mypic {  
-    width:9vh;
-    height: 8vh;
-    border-radius: 50%;
-}
-
-.recpic {
-    width:6vh;
-    height: 5vh;
-    border-radius: 50%;
-}
-
-.recs {
-    position: fixed;
-    right: 1%;
-    padding: 1%;
-    height: 90%;
-    width: 32%;
-}
-
-.recheader {
-    width: 100%;
-    font-family: 'Comfortaa', cursive;
-}
-
-.suggestionsforyou{
-    font-weight: bolder;
-    color: rgb(88, 88, 88);
-    padding-bottom: 2%;
-}
-
-.toprec {
-    padding-top: 3%;;
-}
-
-.midrec {
-    flex-direction: column;
-    height: 5vh;
-    width: 98%;  
-}
-
-.recname {
-    font-weight: bolder;
-}
-
-.suggested {
-    font-size: 70%;
-    width: 100%;
-}
-
-.follow {
-    justify-content: flex-end;
-    color: blue;
-}
-
-.cards {
-    display: flex;
-    flex-direction: column;
-    padding:1%;
-    background-color: white;
-}
-
-.cards * {
-    /* border: solid 1px #000; */
-}
-
-.top {
-    width: 100%;
-    flex-direction: column;
-}
-
-.hallname {
-    padding-left: 3%;
-    width: 100%;
-    font-family: 'Comfortaa', cursive;
-}
-
-.hallpic {
-    width:100%;
-    height: 60%;
-}
-
-.middle {
-    width: 100%;
-}
-
-.middleleft {
-    flex-direction: column;
-    padding-left: 3%;
-    width: 70%;
-}
-
-.middleright {
-    justify-content: flex-end;
-    width: 30%;
-    padding-right: 3%;
-}
-
-.featuredlabel {
-    font-size: 65%; 
-    font-weight: bolder;
-    background: #452c63;
-    color: white;
-}
-
-.featureditems {
-    flex-direction: column;    
-}
-
-.item1  {
-    padding-top: 1%;;
-    padding-left: 3%;
-    padding-right: 3%;
-    font-weight: bolder;
-}
-
-.item2 {
-    padding-top: 1%;
-    padding-left: 3%;
-    padding-right: 3%;
-    font-weight: bolder;
-}
-
-.waitlabel {
-    font-size: 3vw;
-    font-weight: bolder;
-    width: 10vw;
-    height: 10vw;
-    border-radius: 50%;
-    background: green;
-    text-align: center;
-    margin: 10px;
-    
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-around;
-
-    /* background: radial-gradient(
-    circle at center,
-    rgb(143, 188, 143) 50%,
-    rgb(143, 188, 143)  50%,
-    rgba(0, 128, 172, 0) 50%
-    ); */
-    /* padding-bottom: 25%; */
-}
-.waitlabel > p {
-    margin: 0;
-}
-
-.waittime {
-    flex-direction: column;    
-}
-
-.waitscore {
-    padding-top: 1%;;
-    padding-left: 3%;
-    padding-right: 3%;
-    font-weight: bolder;
-}
-
-.more {
-    padding-top: 1%;
-    padding-left: 3%;
-    padding-right: 3%;
-    font-weight: bolder;
-}
-
-.bottom {
-    padding-left: 3%;
-    padding-top: 1%;
-}
-
-.form {
-    font-family: sans-serif;
-    background: white;
-    padding: 1%;
-}
-
-a {
-    text-decoration: none;
-    color: blue;
-  }
-
-@media (min-width:900px) {
-    .cards {
-        width:100%;
-    }
-}
