+++ src/styles/DiningCard.css
@@ -1,0 +1,128 @@
+.cards {
+  display: flex;
+  flex-direction: column;
+  padding: 1%;
+  background-color: white;
+}
+
+.cards * {
+  /* border: solid 1px #000; */
+}
+
+.top {
+  width: 100%;
+  flex-direction: column;
+}
+
+.hallname {
+  padding-left: 3%;
+  width: 100%;
+  font-family: 'Comfortaa', cursive;
+}
+
+.hallpic {
+  width: 100%;
+  height: 60%;
+}
+
+.middle {
+  width: 100%;
+}
+
+.middleleft {
+  flex-direction: column;
+  padding-left: 3%;
+  width: 70%;
+}
+
+.middleright {
+  justify-content: flex-end;
+  width: 30%;
+  padding-right: 3%;
+}
+
+.featuredlabel {
+  font-size: 65%;
+  font-weight: bolder;
+  background: #452c63;
+  color: white;
+}
+
+.featureditems {
+  flex-direction: column;
+}
+
+.item1 {
+  padding-top: 1%;
+  padding-left: 3%;
+  padding-right: 3%;
+  font-weight: bolder;
+}
+
+.item2 {
+  padding-top: 1%;
+  padding-left: 3%;
+  padding-right: 3%;
+  font-weight: bolder;
+}
+
+.waitlabel {
+  font-size: 3vw;
+  font-weight: bolder;
+  width: 10vw;
+  height: 10vw;
+  border-radius: 50%;
+  background: green;
+  text-align: center;
+  margin: 10px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-around;
+
+  /* background: radial-gradient(
+    circle at center,
+    rgb(143, 188, 143) 50%,
+    rgb(143, 188, 143)  50%,
+    rgba(0, 128, 172, 0) 50%
+    ); */
+  /* padding-bottom: 25%; */
+}
+.waitlabel > p {
+  margin: 0;
+}
+
+.waittime {
+  flex-direction: column;
+}
+
+.waitscore {
+  padding-top: 1%;
+  padding-left: 3%;
+  padding-right: 3%;
+  font-weight: bolder;
+}
+
+.more {
+  padding-top: 1%;
+  padding-left: 3%;
+  padding-right: 3%;
+  font-weight: bolder;
+}
+
+.bottom {
+  padding-left: 3%;
+  padding-top: 1%;
+}
+
+a {
+  text-decoration: none;
+  color: blue;
+}
+
+@media (min-width: 900px) {
+  .cards {
+    width: 100%;
+  }
+}
