+++ src/styles/Form.css
@@ -1,0 +1,5 @@
+.form {
+  font-family: sans-serif;
+  background: white;
+  padding: 1%;
+}
